Unify the integration tests of Scalar DB on X
