+++ src/integration-test/java/com/scalar/db/storage/MetadataIntegrationTestBase.java
@@ -1,0 +1,111 @@
+package com.scalar.db.storage;
+
+import com.scalar.db.api.Scan;
+import com.scalar.db.storage.common.metadata.DataType;
+import com.scalar.db.storage.common.metadata.TableMetadata;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.Iterator;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public abstract class MetadataIntegrationTestBase {
+
+  protected static final String NAMESPACE = "integration_testing";
+  protected static final String TABLE = "test_table";
+  protected static final String COL_NAME1 = "c1";
+  protected static final String COL_NAME2 = "c2";
+  protected static final String COL_NAME3 = "c3";
+  protected static final String COL_NAME4 = "c4";
+  protected static final String COL_NAME5 = "c5";
+  protected static final String COL_NAME6 = "c6";
+  protected static final String COL_NAME7 = "c7";
+  protected static final String COL_NAME8 = "c8";
+  protected static final String COL_NAME9 = "c9";
+  protected static final String COL_NAME10 = "c10";
+  protected static final String COL_NAME11 = "c11";
+
+  protected TableMetadata tableMetadata;
+
+  protected abstract TableMetadata getTableMetadata();
+
+  @Before
+  public void setUp() throws Exception {
+    tableMetadata = getTableMetadata();
+  }
+
+  @Test
+  public void testMetadataIsNotNull() {
+    assertThat(tableMetadata).isNotNull();
+  }
+
+  @Test
+  public void testPartitionKeyNames() {
+    assertThat(tableMetadata.getPartitionKeyNames().size()).isEqualTo(2);
+    Iterator<String> iterator = tableMetadata.getPartitionKeyNames().iterator();
+    assertThat(iterator.next()).isEqualTo(COL_NAME2);
+    assertThat(iterator.next()).isEqualTo(COL_NAME1);
+  }
+
+  @Test
+  public void testClusteringKeyNames() {
+    assertThat(tableMetadata.getClusteringKeyNames().size()).isEqualTo(2);
+    Iterator<String> iterator = tableMetadata.getClusteringKeyNames().iterator();
+    assertThat(iterator.next()).isEqualTo(COL_NAME4);
+    assertThat(iterator.next()).isEqualTo(COL_NAME3);
+  }
+
+  @Test
+  public void testColumnNames() {
+    assertThat(tableMetadata.getColumnNames().size()).isEqualTo(11);
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME1)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME2)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME3)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME4)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME5)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME6)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME7)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME8)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME9)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME10)).isTrue();
+    assertThat(tableMetadata.getColumnNames().contains(COL_NAME11)).isTrue();
+  }
+
+  @Test
+  public void testColumnDataType() {
+    assertThat(tableMetadata.getColumnDataType(COL_NAME1)).isEqualTo(DataType.INT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME2)).isEqualTo(DataType.TEXT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME3)).isEqualTo(DataType.TEXT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME4)).isEqualTo(DataType.INT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME5)).isEqualTo(DataType.INT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME6)).isEqualTo(DataType.TEXT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME7)).isEqualTo(DataType.BIGINT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME8)).isEqualTo(DataType.FLOAT);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME9)).isEqualTo(DataType.DOUBLE);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME10)).isEqualTo(DataType.BOOLEAN);
+    assertThat(tableMetadata.getColumnDataType(COL_NAME11)).isEqualTo(DataType.BLOB);
+  }
+
+  @Test
+  public void testClusteringOrder() {
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME1)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME2)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME3)).isEqualTo(Scan.Ordering.Order.DESC);
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME4)).isEqualTo(Scan.Ordering.Order.ASC);
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME5)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME6)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME7)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME8)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME9)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME10)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME11)).isNull();
+  }
+
+  @Test
+  public void testSecondaryIndexNames() {
+    assertThat(tableMetadata.getSecondaryIndexNames().size()).isEqualTo(2);
+    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME5)).isTrue();
+    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME6)).isTrue();
+  }
+}
