+++ src/integration-test/java/com/scalar/db/storage/dynamo/DynamoMetadataIntegrationTest.java
@@ -3,9 +3,10 @@ package com.scalar.db.storage.dynamo;
 import com.scalar.db.api.Get;
 import com.scalar.db.api.Scan;
 import com.scalar.db.io.Key;
-import com.scalar.db.storage.common.metadata.DataType;
-import org.junit.After;
-import org.junit.Before;
+import com.scalar.db.storage.MetadataIntegrationTestBase;
+import com.scalar.db.storage.common.metadata.TableMetadata;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
@@ -25,36 +26,46 @@ import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType;
 import java.net.URI;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class DynamoMetadataIntegrationTest {
+public class DynamoMetadataIntegrationTest extends MetadataIntegrationTestBase {
 
   private static final String METADATA_DATABASE = "scalardb";
   private static final String METADATA_TABLE = "metadata";
-  private static final String DATABASE = "integration_testing";
-  private static final String TABLE = "test_table";
 
-  private static final String COL_NAME1 = "c1";
-  private static final String COL_NAME2 = "c2";
-  private static final String COL_NAME3 = "c3";
-  private static final String COL_NAME4 = "c4";
-  private static final String COL_NAME5 = "c5";
-  private static final String COL_NAME6 = "c6";
-  private static final String COL_NAME7 = "c7";
-  private static final String COL_NAME8 = "c8";
-  private static final String COL_NAME9 = "c9";
-  private static final String COL_NAME10 = "c10";
-  private static final String COL_NAME11 = "c11";
+  private static Optional<String> namespacePrefix;
+  private static DynamoDbClient client;
+  private static TableMetadataManager tableMetadataManager;
 
-  private Optional<String> namespacePrefix;
-  private DynamoDbClient client;
+  @Override
+  protected TableMetadata getTableMetadata() {
+    Get dummyOperation = new Get(new Key()).forNamespace(NAMESPACE).forTable(TABLE);
+    namespacePrefix.ifPresent(n -> dummyOperation.forNamespacePrefix(namespacePrefix().get()));
+    return tableMetadataManager.getTableMetadata(dummyOperation);
+  }
 
-  @Before
-  public void setUp() throws Exception {
+  @Test
+  @Override
+  public void testClusteringOrder() {
+    // Fow now, the clustering order is always ASC in the DynamoDB adapter
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME1)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME2)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME3)).isEqualTo(Scan.Ordering.Order.ASC);
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME4)).isEqualTo(Scan.Ordering.Order.ASC);
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME5)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME6)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME7)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME8)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME9)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME10)).isNull();
+    assertThat(tableMetadata.getClusteringOrder(COL_NAME11)).isNull();
+  }
+
+  @BeforeClass
+  public static void setUpBeforeClass() throws Exception {
     String endpointOverride =
         System.getProperty("scalardb.dynamo.endpoint_override", "http://localhost:8000");
     String region = System.getProperty("scalardb.dynamo.region", "us-west-2");
@@ -106,7 +117,7 @@ public class DynamoMetadataIntegrationTest {
 
     // insert metadata
     Map<String, AttributeValue> values = new HashMap<>();
-    values.put("table", AttributeValue.builder().s(table(DATABASE, TABLE)).build());
+    values.put("table", AttributeValue.builder().s(table(NAMESPACE, TABLE)).build());
     values.put(
         "partitionKey",
         AttributeValue.builder()
@@ -146,87 +157,23 @@ public class DynamoMetadataIntegrationTest {
             .build();
 
     client.putItem(putItemRequest);
+
+    tableMetadataManager = new TableMetadataManager(client, namespacePrefix());
   }
 
-  @After
-  public void tearDown() throws Exception {
+  @AfterClass
+  public static void tearDownAfterClass() throws Exception {
     client.deleteTable(
         DeleteTableRequest.builder().tableName(table(METADATA_DATABASE, METADATA_TABLE)).build());
 
     client.close();
   }
 
-  private Optional<String> namespacePrefix() {
+  private static Optional<String> namespacePrefix() {
     return namespacePrefix.map(n -> n + "_");
   }
 
-  private String database(String database) {
-    return namespacePrefix().orElse("") + database;
-  }
-
-  private String table(String database, String table) {
-    return database(database) + "." + table;
-  }
-
-  @Test
-  public void test() throws Exception {
-    TableMetadataManager tableMetadataManager = new TableMetadataManager(client, namespacePrefix());
-
-    Get dummyOperation = new Get(new Key()).forNamespace(DATABASE).forTable(TABLE);
-    namespacePrefix.ifPresent(n -> dummyOperation.forNamespacePrefix(namespacePrefix().get()));
-    DynamoTableMetadata tableMetadata = tableMetadataManager.getTableMetadata(dummyOperation);
-
-    assertThat(tableMetadata).isNotNull();
-
-    assertThat(tableMetadata.getPartitionKeyNames().size()).isEqualTo(2);
-    Iterator<String> iterator = tableMetadata.getPartitionKeyNames().iterator();
-    assertThat(iterator.next()).isEqualTo(COL_NAME2);
-    assertThat(iterator.next()).isEqualTo(COL_NAME1);
-
-    assertThat(tableMetadata.getClusteringKeyNames().size()).isEqualTo(2);
-    iterator = tableMetadata.getClusteringKeyNames().iterator();
-    assertThat(iterator.next()).isEqualTo(COL_NAME4);
-    assertThat(iterator.next()).isEqualTo(COL_NAME3);
-
-    assertThat(tableMetadata.getColumnNames().size()).isEqualTo(11);
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME1)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME2)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME3)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME4)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME5)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME6)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME7)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME8)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME9)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME10)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME11)).isTrue();
-
-    assertThat(tableMetadata.getColumnDataType(COL_NAME1)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME2)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME3)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME4)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME5)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME6)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME7)).isEqualTo(DataType.BIGINT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME8)).isEqualTo(DataType.FLOAT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME9)).isEqualTo(DataType.DOUBLE);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME10)).isEqualTo(DataType.BOOLEAN);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME11)).isEqualTo(DataType.BLOB);
-
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME1)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME2)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME3)).isEqualTo(Scan.Ordering.Order.ASC);
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME4)).isEqualTo(Scan.Ordering.Order.ASC);
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME5)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME6)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME7)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME8)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME9)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME10)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME11)).isNull();
-
-    assertThat(tableMetadata.getSecondaryIndexNames().size()).isEqualTo(2);
-    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME5)).isTrue();
-    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME6)).isTrue();
+  private static String table(String database, String table) {
+    return namespacePrefix().orElse("") + database + "." + table;
   }
 }
