+++ src/integration-test/java/com/scalar/db/storage/jdbc/JdbcMetadataIntegrationTest.java
@@ -1,41 +1,71 @@
 package com.scalar.db.storage.jdbc;
 
 import com.scalar.db.api.Scan;
+import com.scalar.db.storage.MetadataIntegrationTestBase;
 import com.scalar.db.storage.common.metadata.DataType;
+import com.scalar.db.storage.common.metadata.TableMetadata;
 import com.scalar.db.storage.jdbc.metadata.JdbcTableMetadata;
 import com.scalar.db.storage.jdbc.metadata.TableMetadataManager;
 import com.scalar.db.storage.jdbc.test.TestEnv;
-import org.junit.After;
-import org.junit.Before;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
+import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class JdbcMetadataIntegrationTest {
+public class JdbcMetadataIntegrationTest extends MetadataIntegrationTestBase {
 
-  private static final String NAMESPACE = "integration_testing";
-  private static final String TABLE = "test_table";
-  private static final String COL_NAME1 = "c1";
-  private static final String COL_NAME2 = "c2";
-  private static final String COL_NAME3 = "c3";
-  private static final String COL_NAME4 = "c4";
-  private static final String COL_NAME5 = "c5";
-  private static final String COL_NAME6 = "c6";
-  private static final String COL_NAME7 = "c7";
-  private static final String COL_NAME8 = "c8";
-  private static final String COL_NAME9 = "c9";
-  private static final String COL_NAME10 = "c10";
-  private static final String COL_NAME11 = "c11";
+  private static TestEnv testEnv;
+  private static TableMetadataManager tableMetadataManager;
 
-  private TestEnv testEnv;
+  @Override
+  protected TableMetadata getTableMetadata() {
+    try {
+      Optional<String> namespacePrefix = testEnv.getJdbcDatabaseConfig().getNamespacePrefix();
+      String fullTableName = namespacePrefix.orElse("") + NAMESPACE + "." + TABLE;
+      return tableMetadataManager.getTableMetadata(fullTableName);
+    } catch (SQLException e) {
+      throw new RuntimeException(e);
+    }
+  }
 
-  @Before
-  public void setUp() throws Exception {
+  @Test
+  public void testSchemaAndTableName() {
+    JdbcTableMetadata jdbcTableMetadata = (JdbcTableMetadata) tableMetadata;
+
+    Optional<String> namespacePrefix = testEnv.getJdbcDatabaseConfig().getNamespacePrefix();
+    String schema = namespacePrefix.orElse("") + NAMESPACE;
+    assertThat(jdbcTableMetadata.getSchema()).isEqualTo(schema);
+    assertThat(jdbcTableMetadata.getTable()).isEqualTo(TABLE);
+    String fullTableName = schema + "." + TABLE;
+    assertThat(jdbcTableMetadata.getFullTableName()).isEqualTo(fullTableName);
+  }
+
+  @Test
+  public void testSecondaryIndexOrder() {
+    JdbcTableMetadata jdbcTableMetadata = (JdbcTableMetadata) tableMetadata;
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME1)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME2)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME3)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME4)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME5))
+        .isEqualTo(Scan.Ordering.Order.ASC);
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME6))
+        .isEqualTo(Scan.Ordering.Order.DESC);
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME7)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME8)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME9)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME10)).isNull();
+    assertThat(jdbcTableMetadata.getSecondaryIndexOrder(COL_NAME11)).isNull();
+  }
+
+  @BeforeClass
+  public static void setUpBeforeClass() throws Exception {
     testEnv = new TestEnv();
     testEnv.register(
         NAMESPACE,
@@ -70,89 +100,17 @@ public class JdbcMetadataIntegrationTest {
             put(COL_NAME6, Scan.Ordering.Order.DESC);
           }
         });
-    testEnv.createTables();
-  }
+    testEnv.createMetadataTable();
+    testEnv.insertMetadata();
 
-  @After
-  public void tearDown() throws Exception {
-    testEnv.dropTables();
-    testEnv.close();
-  }
-
-  @Test
-  public void test() throws Exception {
     Optional<String> namespacePrefix = testEnv.getJdbcDatabaseConfig().getNamespacePrefix();
-    TableMetadataManager tableMetadataManager =
+    tableMetadataManager =
         new TableMetadataManager(testEnv.getDataSource(), namespacePrefix, testEnv.getRdbEngine());
+  }
 
-    String fullTableName = namespacePrefix.orElse("") + NAMESPACE + "." + TABLE;
-    JdbcTableMetadata tableMetadata = tableMetadataManager.getTableMetadata(fullTableName);
-
-    assertThat(tableMetadata).isNotNull();
-
-    assertThat(tableMetadata.getFullTableName()).isEqualTo(fullTableName);
-
-    assertThat(tableMetadata.getPartitionKeyNames().size()).isEqualTo(2);
-    Iterator<String> iterator = tableMetadata.getPartitionKeyNames().iterator();
-    assertThat(iterator.next()).isEqualTo(COL_NAME2);
-    assertThat(iterator.next()).isEqualTo(COL_NAME1);
-
-    assertThat(tableMetadata.getClusteringKeyNames().size()).isEqualTo(2);
-    iterator = tableMetadata.getClusteringKeyNames().iterator();
-    assertThat(iterator.next()).isEqualTo(COL_NAME4);
-    assertThat(iterator.next()).isEqualTo(COL_NAME3);
-
-    assertThat(tableMetadata.getColumnNames().size()).isEqualTo(11);
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME1)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME2)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME3)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME4)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME5)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME6)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME7)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME8)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME9)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME10)).isTrue();
-    assertThat(tableMetadata.getColumnNames().contains(COL_NAME11)).isTrue();
-
-    assertThat(tableMetadata.getColumnDataType(COL_NAME1)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME2)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME3)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME4)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME5)).isEqualTo(DataType.INT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME6)).isEqualTo(DataType.TEXT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME7)).isEqualTo(DataType.BIGINT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME8)).isEqualTo(DataType.FLOAT);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME9)).isEqualTo(DataType.DOUBLE);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME10)).isEqualTo(DataType.BOOLEAN);
-    assertThat(tableMetadata.getColumnDataType(COL_NAME11)).isEqualTo(DataType.BLOB);
-
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME1)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME2)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME3)).isEqualTo(Scan.Ordering.Order.DESC);
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME4)).isEqualTo(Scan.Ordering.Order.ASC);
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME5)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME6)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME7)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME8)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME9)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME10)).isNull();
-    assertThat(tableMetadata.getClusteringOrder(COL_NAME11)).isNull();
-
-    assertThat(tableMetadata.getSecondaryIndexNames().size()).isEqualTo(2);
-    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME5)).isTrue();
-    assertThat(tableMetadata.getSecondaryIndexNames().contains(COL_NAME6)).isTrue();
-
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME1)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME2)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME3)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME4)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME5)).isEqualTo(Scan.Ordering.Order.ASC);
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME6)).isEqualTo(Scan.Ordering.Order.DESC);
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME7)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME8)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME9)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME10)).isNull();
-    assertThat(tableMetadata.getSecondaryIndexOrder(COL_NAME11)).isNull();
+  @AfterClass
+  public static void tearDownAfterClass() throws Exception {
+    testEnv.dropMetadataTable();
+    testEnv.close();
   }
 }
