+++ src/integration-test/java/com/scalar/db/storage/jdbc/test/TestEnv.java
@@ -293,6 +293,12 @@ public class TestEnv implements Closeable {
     }
   }
 
+  public void insertMetadata() throws SQLException {
+    for (JdbcTableMetadata metadata : metadataList) {
+      insertMetadata(metadata);
+    }
+  }
+
   private void insertMetadata(JdbcTableMetadata metadata) throws SQLException {
     int ordinalPosition = 1;
     for (String partitionKeyName : metadata.getPartitionKeyNames()) {
@@ -352,7 +358,7 @@ public class TestEnv implements Closeable {
     execute("DROP TABLE " + enclosedFullTableName(metadata.getSchema(), metadata.getTable()));
   }
 
-  private void createMetadataTable() throws SQLException {
+  public void createMetadataTable() throws SQLException {
     createSchema(getMetadataSchema());
 
     // create the metadata table
@@ -396,7 +402,7 @@ public class TestEnv implements Closeable {
     }
   }
 
-  private void dropMetadataTable() throws SQLException {
+  public void dropMetadataTable() throws SQLException {
     // drop the metadata table
     execute("DROP TABLE " + enclosedMetadataTableName());
 
@@ -404,8 +410,6 @@ public class TestEnv implements Closeable {
   }
 
   public void createTables() throws SQLException {
-    createMetadataTable();
-
     Set<String> schemas =
         metadataList.stream().map(JdbcTableMetadata::getSchema).collect(Collectors.toSet());
     for (String schema : schemas) {
@@ -419,15 +423,19 @@ public class TestEnv implements Closeable {
     for (JdbcTableMetadata metadata : metadataList) {
       createIndex(metadata);
     }
+  }
 
+  public void deleteTableData() throws SQLException {
     for (JdbcTableMetadata metadata : metadataList) {
-      insertMetadata(metadata);
+      deleteTableData(metadata);
     }
   }
 
-  public void dropTables() throws SQLException {
-    dropMetadataTable();
+  private void deleteTableData(JdbcTableMetadata metadata) throws SQLException {
+    execute("DELETE FROM " + enclosedFullTableName(metadata.getSchema(), metadata.getTable()));
+  }
 
+  public void dropTables() throws SQLException {
     for (JdbcTableMetadata metadata : metadataList) {
       dropTable(metadata);
     }
