+++ src/integration-test/java/com/scalar/db/transaction/consensuscommit/ConsensusCommitWithCosmosIntegrationTest.java
@@ -19,7 +19,6 @@ import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Test;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -32,12 +31,10 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
 
-import static com.scalar.db.transaction.consensuscommit.ConsensusCommitIntegrationTest.NAMESPACE;
-import static com.scalar.db.transaction.consensuscommit.ConsensusCommitIntegrationTest.TABLE_1;
-import static com.scalar.db.transaction.consensuscommit.ConsensusCommitIntegrationTest.TABLE_2;
 import static org.mockito.Mockito.spy;
 
-public class ConsensusCommitWithCosmosIntegrationTest {
+public class ConsensusCommitWithCosmosIntegrationTest extends ConsensusCommitIntegrationTestBase {
+
   private static final String METADATA_DATABASE = "scalardb";
   private static final String METADATA_CONTAINER = "metadata";
   private static final String STORED_PROCEDURE_PATH =
@@ -47,8 +44,7 @@ public class ConsensusCommitWithCosmosIntegrationTest {
   private static Optional<String> namespacePrefix;
   private static CosmosClient client;
   private static DistributedStorage originalStorage;
-
-  private ConsensusCommitIntegrationTest test;
+  private static DatabaseConfig config;
 
   @Before
   public void setUp() throws IOException {
@@ -56,17 +52,9 @@ public class ConsensusCommitWithCosmosIntegrationTest {
     Coordinator coordinator = spy(new Coordinator(storage));
     RecoveryHandler recovery = spy(new RecoveryHandler(storage, coordinator));
     CommitHandler commit = spy(new CommitHandler(storage, coordinator, recovery));
-
-    Properties props = new Properties();
-    props.setProperty(DatabaseConfig.CONTACT_POINTS, "dummy");
-    props.setProperty(DatabaseConfig.USERNAME, "dummy");
-    props.setProperty(DatabaseConfig.PASSWORD, "dummy");
     ConsensusCommitManager manager =
-        new ConsensusCommitManager(
-            storage, new DatabaseConfig(props), coordinator, recovery, commit);
-
-    test = new ConsensusCommitIntegrationTest(manager, storage, coordinator, recovery);
-    test.setUp();
+        new ConsensusCommitManager(storage, config, coordinator, recovery, commit);
+    setUp(manager, storage, coordinator, recovery);
   }
 
   @After
@@ -91,403 +79,6 @@ public class ConsensusCommitWithCosmosIntegrationTest {
           .getContainer(container)
           .deleteItem(record, new CosmosItemRequestOptions());
     }
-  }
-
-  @Test
-  public void get_GetGivenForCommittedRecord_ShouldReturnRecord() throws Exception {
-    test.get_GetGivenForCommittedRecord_ShouldReturnRecord();
-  }
-
-  @Test
-  public void scan_ScanGivenForCommittedRecord_ShouldReturnRecord() throws Exception {
-    test.scan_ScanGivenForCommittedRecord_ShouldReturnRecord();
-  }
-
-  @Test
-  public void get_CalledTwice_ShouldReturnFromSnapshotInSecondTime() throws Exception {
-    test.get_CalledTwice_ShouldReturnFromSnapshotInSecondTime();
-  }
-
-  @Test
-  public void
-      get_CalledTwiceAndAnotherTransactionCommitsInBetween_ShouldReturnFromSnapshotInSecondTime()
-          throws Exception {
-    test
-        .get_CalledTwiceAndAnotherTransactionCommitsInBetween_ShouldReturnFromSnapshotInSecondTime();
-  }
-
-  @Test
-  public void get_GetGivenForNonExisting_ShouldReturnEmpty() throws Exception {
-    test.get_GetGivenForNonExisting_ShouldReturnEmpty();
-  }
-
-  @Test
-  public void scan_ScanGivenForNonExisting_ShouldReturnEmpty() throws Exception {
-    test.scan_ScanGivenForNonExisting_ShouldReturnEmpty();
-  }
-
-  @Test
-  public void get_GetGivenForPreparedWhenCoordinatorStateCommitted_ShouldRollforward()
-      throws Exception {
-    test.get_GetGivenForPreparedWhenCoordinatorStateCommitted_ShouldRollforward();
-  }
-
-  @Test
-  public void scan_ScanGivenForPreparedWhenCoordinatorStateCommitted_ShouldRollforward()
-      throws Exception {
-    test.scan_ScanGivenForPreparedWhenCoordinatorStateCommitted_ShouldRollforward();
-  }
-
-  @Test
-  public void get_GetGivenForPreparedWhenCoordinatorStateAborted_ShouldRollback() throws Exception {
-    test.get_GetGivenForPreparedWhenCoordinatorStateAborted_ShouldRollback();
-  }
-
-  @Test
-  public void scan_ScanGivenForPreparedWhenCoordinatorStateAborted_ShouldRollback()
-      throws Exception {
-    test.scan_ScanGivenForPreparedWhenCoordinatorStateAborted_ShouldRollback();
-  }
-
-  @Test
-  public void
-      get_GetGivenForPreparedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction()
-          throws Exception {
-    test
-        .get_GetGivenForPreparedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForPreparedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction()
-          throws Exception {
-    test
-        .scan_ScanGivenForPreparedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction();
-  }
-
-  @Test
-  public void get_GetGivenForPreparedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction()
-      throws Exception {
-    test.get_GetGivenForPreparedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForPreparedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction()
-          throws Exception {
-    test.scan_ScanGivenForPreparedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction();
-  }
-
-  @Test
-  public void
-      get_GetGivenForPreparedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly()
-          throws Exception {
-    test
-        .get_GetGivenForPreparedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForPreparedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly()
-          throws Exception {
-    test
-        .scan_ScanGivenForPreparedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly();
-  }
-
-  @Test
-  public void
-      get_GetGivenForPreparedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly()
-          throws Exception {
-    test
-        .get_GetGivenForPreparedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForPreparedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly()
-          throws Exception {
-    test
-        .scan_ScanGivenForPreparedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly();
-  }
-
-  @Test
-  public void get_GetGivenForDeletedWhenCoordinatorStateCommitted_ShouldRollforward()
-      throws Exception {
-    test.get_GetGivenForDeletedWhenCoordinatorStateCommitted_ShouldRollforward();
-  }
-
-  @Test
-  public void scan_ScanGivenForDeletedWhenCoordinatorStateCommitted_ShouldRollforward()
-      throws Exception {
-    test.scan_ScanGivenForDeletedWhenCoordinatorStateCommitted_ShouldRollforward();
-  }
-
-  @Test
-  public void get_GetGivenForDeletedWhenCoordinatorStateAborted_ShouldRollback() throws Exception {
-    test.get_GetGivenForDeletedWhenCoordinatorStateAborted_ShouldRollback();
-  }
-
-  @Test
-  public void scan_ScanGivenForDeletedWhenCoordinatorStateAborted_ShouldRollback()
-      throws Exception {
-    test.scan_ScanGivenForDeletedWhenCoordinatorStateAborted_ShouldRollback();
-  }
-
-  @Test
-  public void
-      get_GetGivenForDeletedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction()
-          throws Exception {
-    test
-        .get_GetGivenForDeletedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForDeletedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction()
-          throws Exception {
-    test
-        .scan_ScanGivenForDeletedWhenCoordinatorStateNotExistAndNotExpired_ShouldNotAbortTransaction();
-  }
-
-  @Test
-  public void get_GetGivenForDeletedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction()
-      throws Exception {
-    test.get_GetGivenForDeletedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForDeletedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction()
-          throws Exception {
-    test.scan_ScanGivenForDeletedWhenCoordinatorStateNotExistAndExpired_ShouldAbortTransaction();
-  }
-
-  @Test
-  public void
-      get_GetGivenForDeletedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly()
-          throws Exception {
-    test
-        .get_GetGivenForDeletedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForDeletedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly()
-          throws Exception {
-    test
-        .scan_ScanGivenForDeletedWhenCoordinatorStateCommittedAndRollforwardedByAnother_ShouldRollforwardProperly();
-  }
-
-  @Test
-  public void
-      get_GetGivenForDeletedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly()
-          throws Exception {
-    test
-        .get_GetGivenForDeletedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly();
-  }
-
-  @Test
-  public void
-      scan_ScanGivenForDeletedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly()
-          throws Exception {
-    test
-        .scan_ScanGivenForDeletedWhenCoordinatorStateAbortedAndRollbackedByAnother_ShouldRollbackProperly();
-  }
-
-  @Test
-  public void getAndScan_CommitHappenedInBetween_ShouldReadRepeatably() throws Exception {
-    test.getAndScan_CommitHappenedInBetween_ShouldReadRepeatably();
-  }
-
-  @Test
-  public void putAndCommit_PutGivenForNonExisting_ShouldCreateRecord() throws Exception {
-    test.putAndCommit_PutGivenForNonExisting_ShouldCreateRecord();
-  }
-
-  @Test
-  public void put_PutGivenForExistingAfterRead_ShouldUpdateRecord() throws Exception {
-    test.putAndCommit_PutGivenForExistingAfterRead_ShouldUpdateRecord();
-  }
-
-  @Test
-  public void put_PutGivenForExistingAndNeverRead_ShouldFailWithException() throws Exception {
-    test.putAndCommit_PutGivenForExistingAndNeverRead_ShouldThrowCommitException();
-  }
-
-  @Test
-  public void put_SinglePartitionMutationsGiven_ShouldAccessStorageOnceForPrepareAndCommit()
-      throws Exception {
-    test.putAndCommit_SinglePartitionMutationsGiven_ShouldAccessStorageOnceForPrepareAndCommit();
-  }
-
-  @Test
-  public void put_TwoPartitionsMutationsGiven_ShouldAccessStorageTwiceForPrepareAndCommit()
-      throws Exception {
-    test.putAndCommit_TwoPartitionsMutationsGiven_ShouldAccessStorageTwiceForPrepareAndCommit();
-  }
-
-  @Test
-  public void putAndCommit_MultipleGetAndPutGiven_ShouldCommitProperly() throws Exception {
-    test.putAndCommit_GetsAndPutsGiven_ShouldCommitProperly();
-  }
-
-  @Test
-  public void commit_ConflictingPutsGivenForNonExisting_ShouldCommitEither() throws Exception {
-    test.commit_ConflictingPutsGivenForNonExisting_ShouldCommitOneAndAbortTheOther();
-  }
-
-  @Test
-  public void commit_ConflictingPutsGivenForExisting_ShouldCommitEither() throws Exception {
-    test.commit_ConflictingPutsGivenForExisting_ShouldCommitOneAndAbortTheOther();
-  }
-
-  @Test
-  public void commit_ConflictingPutAndDeleteGivenForExisting_ShouldCommitPutAndAbortDelete()
-      throws Exception {
-    test.commit_ConflictingPutAndDeleteGivenForExisting_ShouldCommitPutAndAbortDelete();
-  }
-
-  @Test
-  public void commit_NonConflictingPutsGivenForExisting_ShouldCommitBoth() throws Exception {
-    test.commit_NonConflictingPutsGivenForExisting_ShouldCommitBoth();
-  }
-
-  @Test
-  public void commit_DeleteGivenWithoutRead_ShouldThrowInvalidUsageException() {
-    test.commit_DeleteGivenWithoutRead_ShouldThrowInvalidUsageException();
-  }
-
-  @Test
-  public void commit_DeleteGivenForNonExisting_ShouldThrowInvalidUsageException() throws Exception {
-    test.commit_DeleteGivenForNonExisting_ShouldThrowInvalidUsageException();
-  }
-
-  @Test
-  public void commit_DeleteGivenForExistingAfterRead_ShouldDeleteRecord() throws Exception {
-    test.commit_DeleteGivenForExistingAfterRead_ShouldDeleteRecord();
-  }
-
-  @Test
-  public void commit_ConflictingDeletesGivenForExisting_ShouldCommitOneAndAbortTheOther()
-      throws Exception {
-    test.commit_ConflictingDeletesGivenForExisting_ShouldCommitOneAndAbortTheOther();
-  }
-
-  @Test
-  public void commit_NonConflictingDeletesGivenForExisting_ShouldCommitBoth() throws Exception {
-    test.commit_NonConflictingDeletesGivenForExisting_ShouldCommitBoth();
-  }
-
-  @Test
-  public void commit_WriteSkewOnExistingRecordsWithSnapshot_ShouldProduceNonSerializableResult()
-      throws Exception {
-    test.commit_WriteSkewOnExistingRecordsWithSnapshot_ShouldProduceNonSerializableResult();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewOnExistingRecordsWithSerializableWithExtraWrite_OneShouldCommitTheOtherShouldThrowCommitConflictException()
-          throws Exception {
-    test
-        .commit_WriteSkewOnExistingRecordsWithSerializableWithExtraWrite_OneShouldCommitTheOtherShouldThrowCommitConflictException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewOnExistingRecordsWithSerializableWithExtraRead_OneShouldCommitTheOtherShouldThrowCommitConflictException()
-          throws Exception {
-    test
-        .commit_WriteSkewOnExistingRecordsWithSerializableWithExtraRead_OneShouldCommitTheOtherShouldThrowCommitConflictException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraWrite_OneShouldCommitTheOtherShouldThrowCommitException()
-          throws Exception {
-    test
-        .commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraWrite_OneShouldCommitTheOtherShouldThrowCommitException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraWriteAndCommitStatusFailed_ShouldRollbackProperly()
-          throws Exception {
-    test
-        .commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraWriteAndCommitStatusFailed_ShouldRollbackProperly();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraRead_OneShouldCommitTheOtherShouldThrowCommitException()
-          throws Exception {
-    test
-        .commit_WriteSkewOnNonExistingRecordsWithSerializableWithExtraRead_OneShouldCommitTheOtherShouldThrowCommitException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewWithScanOnNonExistingRecordsWithSerializableWithExtraWrite_ShouldThrowCommitException()
-          throws Exception {
-    test
-        .commit_WriteSkewWithScanOnNonExistingRecordsWithSerializableWithExtraWrite_ShouldThrowCommitException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewWithScanOnNonExistingRecordsWithSerializableWithExtraRead_ShouldThrowCommitException()
-          throws Exception {
-    test
-        .commit_WriteSkewWithScanOnNonExistingRecordsWithSerializableWithExtraRead_ShouldThrowCommitException();
-  }
-
-  @Test
-  public void
-      commit_WriteSkewWithScanOnExistingRecordsWithSerializableWithExtraRead_ShouldThrowCommitException()
-          throws Exception {
-    test
-        .commit_WriteSkewWithScanOnExistingRecordsWithSerializableWithExtraRead_ShouldThrowCommitException();
-  }
-
-  @Test
-  public void putAndCommit_DeleteGivenInBetweenTransactions_ShouldProduceSerializableResults()
-      throws Exception {
-    test.putAndCommit_DeleteGivenInBetweenTransactions_ShouldProduceSerializableResults();
-  }
-
-  @Test
-  public void deleteAndCommit_DeleteGivenInBetweenTransactions_ShouldProduceSerializableResults()
-      throws Exception {
-    test.deleteAndCommit_DeleteGivenInBetweenTransactions_ShouldProduceSerializableResults();
-  }
-
-  @Test
-  public void get_DeleteCalledBefore_ShouldReturnEmpty() throws Exception {
-    test.get_DeleteCalledBefore_ShouldReturnEmpty();
-  }
-
-  @Test
-  public void scan_DeleteCalledBefore_ShouldReturnEmpty() throws Exception {
-    test.scan_DeleteCalledBefore_ShouldReturnEmpty();
-  }
-
-  @Test
-  public void delete_PutCalledBefore_ShouldDelete() throws Exception {
-    test.delete_PutCalledBefore_ShouldDelete();
-  }
-
-  @Test
-  public void put_DeleteCalledBefore_ShouldPut() throws Exception {
-    test.put_DeleteCalledBefore_ShouldPut();
-  }
-
-  @Test
-  public void scan_OverlappingPutGivenBefore_ShouldThrowCrudRuntimeException() throws Exception {
-    test.scan_OverlappingPutGivenBefore_ShouldThrowCrudRuntimeException();
-  }
-
-  @Test
-  public void scan_NonOverlappingPutGivenBefore_ShouldScan() throws Exception {
-    test.scan_NonOverlappingPutGivenBefore_ShouldScan();
   }
 
   private static String namespacePrefix() {
@@ -574,20 +165,20 @@ public class ConsensusCommitWithCosmosIntegrationTest {
       metadata = new CosmosTableMetadata();
       metadata.setId(table(NAMESPACE, table));
       metadata.setPartitionKeyNames(
-          Collections.singletonList(ConsensusCommitIntegrationTest.ACCOUNT_ID));
+          Collections.singletonList(ConsensusCommitIntegrationTestBase.ACCOUNT_ID));
       metadata.setClusteringKeyNames(
-          Collections.singletonList(ConsensusCommitIntegrationTest.ACCOUNT_TYPE));
+          Collections.singletonList(ConsensusCommitIntegrationTestBase.ACCOUNT_TYPE));
       metadata.setSecondaryIndexNames(Collections.emptySet());
       columns = new HashMap<>();
-      columns.put(ConsensusCommitIntegrationTest.ACCOUNT_ID, "int");
-      columns.put(ConsensusCommitIntegrationTest.ACCOUNT_TYPE, "int");
-      columns.put(ConsensusCommitIntegrationTest.BALANCE, "int");
+      columns.put(ConsensusCommitIntegrationTestBase.ACCOUNT_ID, "int");
+      columns.put(ConsensusCommitIntegrationTestBase.ACCOUNT_TYPE, "int");
+      columns.put(ConsensusCommitIntegrationTestBase.BALANCE, "int");
       columns.put(Attribute.ID, "text");
       columns.put(Attribute.VERSION, "int");
       columns.put(Attribute.STATE, "int");
       columns.put(Attribute.PREPARED_AT, "bigint");
       columns.put(Attribute.COMMITTED_AT, "bigint");
-      columns.put(Attribute.BEFORE_PREFIX + ConsensusCommitIntegrationTest.BALANCE, "int");
+      columns.put(Attribute.BEFORE_PREFIX + ConsensusCommitIntegrationTestBase.BALANCE, "int");
       columns.put(Attribute.BEFORE_ID, "text");
       columns.put(Attribute.BEFORE_VERSION, "int");
       columns.put(Attribute.BEFORE_STATE, "int");
@@ -605,7 +196,7 @@ public class ConsensusCommitWithCosmosIntegrationTest {
     props.setProperty(DatabaseConfig.USERNAME, username);
     props.setProperty(DatabaseConfig.PASSWORD, password);
     namespacePrefix.ifPresent(n -> props.setProperty(DatabaseConfig.NAMESPACE_PREFIX, n));
-    DatabaseConfig config = new DatabaseConfig(props);
+    config = new DatabaseConfig(props);
     originalStorage = new Cosmos(config);
   }
